Remove Int32 convenience methods, as they are misleading

+++ src/main/java/com/basho/riak/client/core/converters/TimeSeriesPBConverter.java
@@ -234,7 +234,7 @@ public final class TimeSeriesPBConverter
         {
             cellBuilder.setBooleanValue(cell.getBoolean());
         }
-        else if(cell.hasInt())
+        else if(cell.hasLong())
         {
             cellBuilder.setIntegerValue(cell.getLong());
         }
